Add GSettings schema overrides for i3-GNOME-Flashback
This commit compiles fixes made by SkyyySi, in setting sane
configuration defaults for GSettings/DConf, as used by GNOME and GNOME
Flashback.

For the most part, the defaults are meant to provide a correct
experience, especially where significant functionality is broken (such
as when a pre-defined desktop background is handled as an invisible
window by i3). Other, more opinionated configuration changes are left to
users.

Co-authored-by: SkyyySi <rasti.b@protonmail.com>